+++ src/events/RegisterTriggersEvent.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace everyday\IncrementalStaticRegeneration\events;
+
+use yii\base\Event;
+
+class RegisterTriggersEvent extends Event
+{
+    /**
+     * @var array
+     *
+     * Additional ISR triggers
+     */
+    public array $triggers;
+}