+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.2.0
+- The ISR request now uses a POST body instead of URL parameters
+- Introduced a `Triggers` service
+- Renamed `additionalEvents` to `additionalTriggers`
+- Fix how we access settings in `helpers/WebRequest`
+
 ## 1.1.0
 - Added support for preview mode (https://nextjs.org/docs/advanced-features/preview-mode)